% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateScale.R
\name{aggregateScale}
\alias{aggregateScale}
\title{Calculate Row-wise Means and Sums for Scales}
\usage{
aggregateScale(dataframe, aggregate = "SUM", NA.threshold = 0.5, na.rm = TRUE)
}
\arguments{
\item{dataframe}{A dataframe containing numeric scores to be aggregated. Each row should correspond to one observation. Each column is one variable.}

\item{aggregate}{Defaults to \code{"SUM"}. Calculates sum scores, use \code{"MEAN"} for mean scores.}

\item{NA.threshold}{The maximum proportion of missingness allowed for one observation/row before the aggregate score is coerced to NA. Defaults to 0.5, which means that rows with NA in more than half of its observation will have an NA aggregate score.}

\item{na.rm}{Whether missing values should be ignored in the calculation for the aggregate scores. Defaults to \code{TRUE}. Does not affect the calculation of proportion of missing data in \code{NA.threshold}.}
}
\value{

}
\description{
Calculate Row-wise Means and Sums for Scales
}
\examples{
mydata <- data.frame(a = c(1:8,NA,NA), b = 1:10, c = rep(NA,10))
aggregate.scale(mydata)
aggregate.scale(mydata, NA.threshold = 0.4)
}
