% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ampliconAnalysis.R
\name{ampliconAnalysis}
\alias{ampliconAnalysis}
\title{ampliconAnalysis}
\usage{
ampliconAnalysis(amplicons, bams, genome, paired = TRUE, minCov = 50,
  min.map.q = 40, mc.cores = 1)
}
\arguments{
  \item{amplicons}{\code{GRanges} of amplicons created by
  \code{ampliconGRanges}}

  \item{bams}{A named \code{character} vector of bams to be
  analysed for the supplied amplicons}

  \item{genome}{\code{BSgenome} to map the amplicons to}

  \item{paired}{\code{boolean} of whether to only load
  properly paired reads from the bams}

  \item{minCov}{Minumum sequencing coverage required to
  report methylation/conversion estimates}

  \item{min.map.q}{Minumum mapping quality for reads to be
  included in the analysis, set to 0 for no filtering}

  \item{mc.cores}{Number of cores to use during processing}
}
\value{
List
}
\description{
Start to finish analysis of MiSeq bisulfite amplicon
sequencing data
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

