% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ampliconGRanges}
\alias{ampliconGRanges}
\title{ampliconGRanges}
\usage{
ampliconGRanges(x, genome, mc.cores = 8)
}
\arguments{
  \item{x}{\code{data.frame} of amplicons containing the
  columns "Amplicon", "Target", "FW", "RV", "Sequenom"}

  \item{genome}{\code{BSgenome} to map the amplicons to}

  \item{mc.cores}{Number of cores to use during mapping}
}
\value{
\code{GRanges} of amplicon locations with metadata
}
\description{
ampliconGRanges
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

