% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotateRegionsByInteractions}
\alias{annotateRegionsByInteractions}
\title{annotateRegionsByInteractions}
\usage{
annotateRegionsByInteractions(reg, tx, interactions, distance = 2000)
}
\arguments{
  \item{reg}{The \code{GRanges} to be annotated}

  \item{tx}{\code{GRanges} of transcripts, annotated with
  \code{gene_name}, \code{gene_id}, \code{tx_id} and
  \code{tx_type}}

  \item{interactions}{\code{GRanges} of interaction data}

  \item{distance}{Distance interaction can be from a
  promoter to count as promoter linked}
}
\value{
\code{reg} with additional per region metadata
}
\description{
Annotate regions with respect to gene annotations using
chromatin interaction data
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

