% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biasPlots.R
\name{biasPlots}
\alias{biasPlots}
\title{biasPlots}
\usage{
biasPlots(x, samples, build)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data - is
  also used to calculate CpG density so should not be
  pre-filtered}

  \item{samples}{\code{data.frame} describing the samples
  to calculate ratios for}

  \item{build}{BSgenome object of the organism the data is
  from}
}
\value{
Called for the side effect of plotting
}
\description{
Plot of sequencing coverage bias versus GC and CpG content
in 100bp genomic bins
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

