% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coverageBarplot}
\alias{coverageBarplot}
\title{coverageBarplot}
\usage{
coverageBarplot(regions, background, annotations, main, nperm = 1000,
  cols = NULL, colBy = c("regions", "annotations"), verbose = FALSE)
}
\arguments{
  \item{regions}{\code{GRangesList} of query regions, for
  example DMRs}

  \item{background}{\code{GRanges} of interrogated genomic
  background}

  \item{annotations}{\code{GRangesList} of subject regions,
  for example ChromHMM annotations}

  \item{main}{Title of the plot}

  \item{nperm}{Number of permutations to perform}

  \item{cols}{Colours to use for the barplot}

  \item{colBy}{Whether to colour by region or annotation}

  \item{verbose}{Whether to print progress}
}
\value{
Creates a barplot of observed/expected ratios of regions vs
annotations with an asterisk to mark those which are
significant, and returns a \code{list} of the plotted data
containing the elements: \item{covTable}{Overlap of
\code{regions} vs \code{annotations} as calculated by
\code{tabulateCoverage}} \item{permTable}{Overlaps of
\code{perms} vs \code{annotations}} \item{permQuant}{2.5\%
and 97.5\% quantiles of observed/expected values obtained
from permuted regions} \item{perms}{Permuted regions used
to calculate p-values} \item{permPval}{P-values for
enrichment and depletion}
}
\description{
Create a barplot of enrichment/depletion of regions vs
annotations, assess significance by permuting regions
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

