% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.diag.R
\name{efa.diag}
\alias{efa.diag}
\title{Calculate KMO and Bartlett' Tests of Sphericity}
\usage{
efa.diag(data)
}
\arguments{
\item{data}{A dataframe containing only the numeric variables to be included in the factor analysis.}
}
\value{

}
\description{
A function that automates and simplifies the code to conduct the Kaiser-Meyer-Olkin (KMO) Test for Sampling Adequacy
and Bartlett's Test of Sphericity. These tests determine the suitability of the data for Factor Analysis.
The code for this function was taken from the notes created by Professor Mike Cheung's statistical module. Visit his website at http://mikewlcheung.github.io/.
}
\examples{
library(psych)
library(lavaan)
efa.diag(HolzingerSwineford1939[,7:15])
}
