% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{extremes}
\alias{extremes}
\title{Identify extreme values based on percentiles}
\usage{
extremes(vec, threshold = 99.5)
}
\arguments{
\item{vec}{Vector}

\item{threshold}{Threshold in percentile expressed between 1 to 100, defaults to 99.5}
}
\value{
Return a logical vector indicating TRUE if value is above the percentile threshold
}
\description{
Identify high improbable values using a percentile threshold
}
\examples{
vec <- c(1:10, 10000)
extremes(vec)
vec[extremes(vec)]
data.frame(vec = vec) \%>\% filter(!extremes(vec))
data.frame(vec = vec) \%>\% filter(extremes(vec))
}
