% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2num.R
\name{fct2num}
\alias{fct2num}
\title{Convert Factor To Numeric While Printing Levels}
\usage{
fct2num(factor_vector)
}
\arguments{
\item{factor_vector}{A vector that is a factor.}
}
\value{

}
\description{
Convert Factor To Numeric While Printing Levels
}
\examples{
a <- factor(c("Agree","Disagree","Agree"))
fct2num(a)
}
