% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findPMDs}
\alias{findPMDs}
\title{findPMDs}
\usage{
findPMDs(x, samples, minCov = 5)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{samples}{\code{data.frame} describing the samples
  to discover PMDs within}

  \item{minCov}{Minumum sequencing coverage for a CpG site
  to contribute to the weighted mean calculation}
}
\value{
\code{GRangesList} of PMDs found in each sample
}
\description{
Find partially domains using the procedure described in
Hovestaft 2014
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

