% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grangesVenn.R
\name{grangesVenn}
\alias{grangesVenn}
\title{grangesVenn}
\usage{
grangesVenn(x, plot = TRUE, basepair = FALSE, ...)
}
\arguments{
  \item{x}{\code{GRangesList} of interest}

  \item{plot}{Whether to plot after calculating the
  intersections}

  \item{basepair}{Whether to calculate the intersections on
  a per-basepair level rather than per region}

  \item{...}{Additional parameters passed on to
  vennDiagram}
}
\value{
A list containing the pooled and reduced ranges, a boolean
matrix of whether each sample in \code{x} overlaps each
\code{Ranges} and the vennCounts
}
\description{
Creates a table of overlapping \code{GRanges} from a
\code{GRangesList}, vennCounts and optionally plots the
first (up to) five
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

