% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadWGBS}
\alias{loadWGBS}
\title{loadWGBS}
\usage{
loadWGBS(path, build)
}
\arguments{
  \item{path}{The path to the project.csv and bigTable.tsv}

  \item{build}{BSgenome object of the organism the data is
  from}
}
\value{
A list with the elements 'samples', 'CpGs' and if a NOMe
experiment 'GCH'
}
\description{
Loads a WGBS experiments bigTable and samples file and
checks consistency
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

