% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/profile.R
\name{lsos}
\alias{lsos}
\title{lsos}
\usage{
lsos(order.by = "Size", decreasing = TRUE, head = TRUE, n = 10)
}
\arguments{
  \item{order.by}{Which object parameter to sort by}

  \item{decreasing}{Sort order}

  \item{head}{Subset list of objects?}

  \item{n}{Number of objects to print if \code{head} is
  \code{TRUE}}
}
\description{
Lists loaded R objects, their dimensions and memory usage
}
\author{
stolen from somewhere on stackoverflow
}

