% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeTx}
\alias{makeTx}
\title{makeTx}
\usage{
makeTx(file, genome, type_attrib = "transcript_biotype")
}
\arguments{
  \item{file}{Filename to uncompressed Ensembl GTF to
  process}

  \item{genome}{\code{BSgenome} object the GTF will be
  associated with}

  \item{type_attrib}{Name of the GTF attribute containing
  the type of transcript - usually "transcript_biotype" or
  "gene_biotype"}
}
\value{
\code{GRanges} of transcripts with appropriate columns for
\code{annotateRegions} namely \code{gene_name},
\code{gene_id}, \code{tx_id} and \code{tx_type}
}
\description{
Create a tx object for annotating regions from an Ensembl
GTF
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

