% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methDensityPlot.R
\name{methDensityPlot}
\alias{methDensityPlot}
\title{methDensityPlot}
\usage{
methDensityPlot(x, samples, title, cols = NULL)
}
\arguments{
  \item{x}{A \code{GRanges}, \code{matrix} or
  \code{data.frame} of methylation ratios}

  \item{samples}{\code{data.frame} describing the samples
  to plot}

  \item{title}{Title for the plot}

  \item{cols}{Named vector of colours to use for each
  group}
}
\value{
Called for the side effect of plotting
}
\description{
Plots the distribution of methylatio ratios in multiple
samples using ggplot
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

