% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methHeatmaps.R
\name{methHeatmap}
\alias{methHeatmap}
\title{methHeatmap}
\usage{
methHeatmap(x, region, samples, CpGislands, tx, extras = list(),
  smoothing = 0, cols = NULL, minCov = 2, ...)
}
\arguments{
  \item{x}{CpG-site level methylation data, including the
  \code{w} column of weights if \code{weighted} is true}

  \item{region}{The genomic region to be plotted}

  \item{samples}{\code{data.frame} describing the samples
  to plot}

  \item{CpGislands}{\code{GRanges} of CpG island positions}

  \item{tx}{\code{GRanges} of transcripts}

  \item{extras}{\code{List} of extra \code{Gviz} tracks to
  plot above the heatmaps}

  \item{smoothing}{If non-zero, the number of basepairs to
  summarise methyation over. If zero no smoothing is
  performed.}

  \item{cols}{Named vector of colours to use for each
  group}

  \item{minCov}{Minimum coverage for a CpG site to be
  included in the weighted mean}

  \item{...}{Additional parameters passed to
  \code{Gviz::plotTracks}}
}
\value{
Create a heatmap of methylation
}
\description{
Plots a region of the genome with per sample heatmaps of
methylation
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

