% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methMDS.R
\name{methMDS}
\alias{methMDS}
\title{methMDS}
\usage{
methMDS(x, samples, smoothing = 1, top = c(1, 10, 100), minCov = 5,
  mc.cores = 1)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{samples}{\code{data.frame} describing the samples
  to calculate ratios for}

  \item{smoothing}{Number of basepairs to smooth
  methylation data over, \code{1} indicates no smoothing}

  \item{top}{\code{numeric vector} of percentage of top
  varying methlation ratios to include in the MDS, defaults
  to the top 1\%, 10\% and 100\%}

  \item{minCov}{minimum amount of coverage to calculate a
  ratio}

  \item{mc.cores}{Number of cores to use}
}
\value{
Called for the side effect of plotting
}
\description{
Plots multidimensional scaling plot of methylation count
data
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

