% Generated by roxygen2 (4.0.1): do not edit by hand
\name{methRatios}
\alias{methRatios}
\title{methRatios}
\usage{
methRatios(x, samples, minCov = 5)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{samples}{\code{data.frame} describing the samples
  to calculate ratios for}

  \item{minCov}{minimum amount of coverage to calculate a
  ratio}
}
\value{
\code{GRanges} of methylation ratios
}
\description{
Calculates methylation ratios from count data
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

