% Generated by roxygen2 (4.0.2): do not edit by hand
\name{methWindowRatios}
\alias{methWindowRatios}
\title{methWindowRatios}
\usage{
methWindowRatios(x, windows, samples, minCov = 5, mc.cores = 1)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{windows}{\code{GRanges} of windows to calculate
  methylation ratios across}

  \item{samples}{\code{data.frame} describing the samples
  to calculate ratios for}

  \item{minCov}{minimum amount of coverage to calculate a
  ratio}

  \item{mc.cores}{Number of cores to use}
}
\value{
\code{GRanges} of methylation ratios
}
\description{
Calculates methylation ratios from count data per each
supplied window
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

