% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NOMe_plots.R
\name{methylationBiPlot}
\alias{methylationBiPlot}
\title{methylationBiPlot}
\usage{
methylationBiPlot(methGpC, methCpG, regions, samples, main = "", up = 1000,
  down = 1000, every = 10, width = 50, minCov = 5, addN = TRUE)
}
\arguments{
  \item{methGpC}{\code{GRanges} of GpC methylation count
  data}

  \item{methCpG}{\code{GRanges} of CpG methylation count
  data}

  \item{regions}{\code{GRanges} to plot methylation and
  occupancy around}

  \item{samples}{\code{data.frame} describing the sample to
  calculate ratios for - must be a single row}

  \item{main}{Title for the plot}

  \item{up}{Distance upstream of regions to plot}

  \item{down}{Distance downstream of regions to plot}

  \item{every}{Resolution to smooth at - distance between
  the start of adjacent smoothing windows}

  \item{width}{Resolution to smooth at - width of each
  smoothing window}

  \item{minCov}{Minimum sequencing coverage for a site to
  be used in calculations}

  \item{addN}{Whether to add the number of regions to the
  title}
}
\value{
\code{GRangesList} of relative position of GpC and CpG
methylation sites to \code{regions} and corresponding
methylation ratio
}
\description{
Plot average methylation and occupancy ratio profiles
around a set of regions for a single sample
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

