% Generated by roxygen2 (4.0.1): do not edit by hand
\name{methylationPlotSamples}
\alias{methylationPlotSamples}
\title{methylationPlotSamples}
\usage{
methylationPlotSamples(meth, regions, samples, GpC, main = "", up = 1000,
  down = 1000, every = 10, width = 50, minCov = 5, addN = TRUE)
}
\arguments{
  \item{meth}{\code{GRanges} of methylation count data
  (either CpG or GpC)}

  \item{regions}{\code{GRanges} to plot
  methylation/occupancy around, one line per element}

  \item{samples}{\code{data.frame} describing the samples
  to calculate ratios for}

  \item{GpC}{Whether the data is GpC data, and therefore
  plotting should be inverted}

  \item{main}{Title for the plot}

  \item{up}{Distance upstream of regions to plot}

  \item{down}{Distance downstream of regions to plot}

  \item{every}{Resolution to smooth at - distance between
  the start of adjacent smoothing windows}

  \item{width}{Resolution to smooth at - width of each
  smoothing window}

  \item{minCov}{Minimum sequencing coverage for a site to
  be used in calculations}

  \item{addN}{Whether to add the number of regions to the
  title}
}
\value{
\code{GRangesList} of per-sample relative position of
methylation sites to \code{regions} and corresponding
methylation ratio
}
\description{
Plot average methylation or occupancy ratio profiles around
a set of regions for multiple samples
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

