% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/never_give_up.R
\name{never_give_up}
\alias{never_give_up}
\title{Wrapper to keep trying an expression until no error is received}
\usage{
never_give_up(expr, sound = "fanfare", message = "Never gonna give you up.")
}
\arguments{
\item{expr}{Expression to be evaluated}

\item{sound}{Fed into beepr::beep(). Defaults to 'fanfare'}
}
\value{
Output of expression when completed
}
\description{
Convenience wrapper to allow an expression to keep evaluating (until 10 attempts) until it completes without error, plays a sound when done if beepr is installed.
}
\examples{
test <- function(e){if(runif(1) < 0.8){stop("Error")}else{"Good"}}
never_give_up(test())
never_give_up(test(), sound = "mario", message = "I wish this would work...")
}
