% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd2prob.R
\name{odd2prob}
\alias{odd2prob}
\title{Convert Odds To Probabilities}
\usage{
odd2prob(odd)
}
\arguments{
\item{p}{Probability value to be converted}
}
\value{
Return the odds
}
\description{
Convert odds to probabilities using the formula of p = odds/(1 + odds)
}
\examples{
odd <- c(0.5,0.75,1)
odd2prob(odd)
}
