% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlapPositions.R
\name{overlapBoundaries}
\alias{overlapBoundaries}
\title{overlapBoundaries}
\usage{
overlapBoundaries(query, subject, distance = 10000, boundary = c("both",
  "start", "end"))
}
\arguments{
  \item{query}{\code{GRanges} of query regions}

  \item{subject}{\code{GRanges} of subject regions}

  \item{distance}{Number of basepairs up and downstream of
  \code{subject} regions to interrogate}

  \item{boundary}{Whether to interrogate the start, end or
  both boundaries of the \code{query} range}
}
\value{
Positions where \code{subject} ranges overlap the
boundaries of \code{query} ranges
}
\description{
Returns
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

