% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overlapMeans}
\alias{overlapMeans}
\title{overlapMeans}
\usage{
overlapMeans(x, y, z, na.rm = FALSE)
}
\arguments{
  \item{x}{The \code{GRanges} object to find overlaps upon}

  \item{y}{The \code{GRanges} object of regions of
  interest}

  \item{z}{The \code{numeric} or \code{integer} vector
  associated with the GRanges \code{x}}

  \item{na.rm}{Logical indicating whether or not to include
  missing values in the results}
}
\value{
A \code{numeric} vector of the same length as \code{y} with
per-range means of \code{z}
}
\description{
Calculate the mean of values associated with ranges which
overlap query ranges
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

