% Generated by roxygen2 (4.0.2): do not edit by hand
\name{overlapPositions}
\alias{overlapPositions}
\title{overlapPositions}
\usage{
overlapPositions(query, subject)
}
\arguments{
  \item{query}{\code{GRanges} of query regions}

  \item{subject}{\code{GRanges} of subject regions (must be
  of width 1 and unstranded)}
}
\value{
Positions where \code{subject} ranges overlap the
\code{query} ranges
}
\description{
Returns the positions where \code{subject} ranges overlap
the \code{query}
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

