% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overlapRatios}
\alias{overlapRatios}
\title{overlapRatios}
\usage{
overlapRatios(x, y, C, cov, minCov = 5)
}
\arguments{
  \item{x}{The \code{GRanges} object to find overlaps upon}

  \item{y}{The \code{GRanges} object of regions of
  interest}

  \item{C}{The numerator vector associated with the GRanges
  \code{x}}

  \item{cov}{The denominator vector associated with the
  GRanges \code{x}}

  \item{minCov}{Minimum per-range coverage required to
  return a ratio}
}
\value{
A \code{numeric} vector of the same length as \code{y} with
per-range ratios of C/cov of \code{z}
}
\description{
Calculate the ratio of C/cov values associated with ranges
which overlap query ranges
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

