% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlapPositions.R
\name{overlapRegionFlanks}
\alias{overlapRegionFlanks}
\title{overlapRegionFlanks}
\usage{
overlapRegionFlanks(query, subject, up = 1000, down = 1000)
}
\arguments{
  \item{query}{\code{GRanges} of query regions}

  \item{subject}{\code{GRanges} of subject regions (must be
  of width 1 and unstranded)}

  \item{up}{Distance upstream to find overlaps}

  \item{down}{Distance downstream to find overlaps}
}
\value{
\code{data.frame} of index and position
}
\description{
Returns the relative position \code{subject} sites around
\code{query} regions \code{start} positions if stranded, or
around \code{query} region centres if unstranded
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

