% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlaps.R
\name{overlapStats}
\alias{overlapStats}
\title{overlapStats}
\usage{
overlapStats(ROI, TFs, TF.gx, gx, accessible = 2.5e+09)
}
\arguments{
  \item{ROI}{\code{GRanges} of interest}

  \item{TFs}{\code{GRangesList} of TF binding sites}

  \item{TF.gx}{Index matching \code{TFs} to rows of
  \code{gx}}

  \item{gx}{\code{GRanges} of gene expression}

  \item{accessible}{Estimate of the number of basepairs of
  the genome that is interrogatable}
}
\value{
A \code{data.frame} with with a row per transcription
factor in \code{TFs} containing the columns: \item{TF}{Name
of the transcription factor} \item{TF.n}{Number of TF
sites} \item{Hits}{Number of \code{ROI} that overlap TF
sites} \item{Hits.percent}{Percent of \code{ROI} that
overlap TF sites} \item{pval}{P-value of significance of
overlap from the hypergeometric distribution}
\item{adj.pval}{Adjusted p-value using "fdr"}
\item{fold}{Estimated fold enrichment of \code{ROI}
overlapping this TF compared to random chance}
\item{Expression}{Expression result from \code{gx}}
\item{RPKM.SA}{Expression result from \code{gx}}
\item{RPKM.VA}{Expression result from \code{gx}}
\item{logFC}{Expression result from \code{gx}}
\item{FDR}{Expression result from \code{gx}}
}
\description{
Calculate statistics of the overlaps of regions of interest
and transcription factor binding sites
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

