% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDensities}
\alias{plotDensities}
\title{plotDensities}
\usage{
plotDensities(s, col = NULL, xlim = NULL, ylim = NULL, na.rm = TRUE,
  main = "", xlab = "", ylab = "Relative Frequency", ...)
}
\arguments{
  \item{s}{A \code{list} or \code{matrix}}

  \item{col}{Colours to use (optional)}

  \item{xlim}{xlim to use, if not supplied are calculated
  from data ranges}

  \item{ylim}{ylim to use, if not supplied are calculated
  from data ranges}

  \item{na.rm}{Whether to remove NA values when calculating
  densites}

  \item{main}{Title for the plot}

  \item{xlab}{xlab to use}

  \item{ylab}{ylab to use}

  \item{...}{Additional parameters passed on to
  \code{plot}}
}
\value{
Called for the side effect of plotting
}
\description{
Plots multiple calls to \code{density} on the same plot
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

