% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotManyUCSC}
\alias{plotManyUCSC}
\title{plotManyUCSC}
\usage{
plotManyUCSC(x, outFile, mc.cores = 1, hub, session = NULL)
}
\arguments{
  \item{x}{\code{GRanges} of regions to screenshot}

  \item{outFile}{Filename of the output pdf}

  \item{mc.cores}{No of threads to use to take screenshots
  - too many will break UCSC/this function}

  \item{hub}{URL of UCSC track hub to load}

  \item{session}{URL of the UCSC session to load
  (optional)}
}
\description{
Takes many UCSC screenshots and collates them together into
a single pdf using the system ghostscript
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

