% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPromoterDMRs}
\alias{plotPromoterDMRs}
\title{plotPromoterDMRs}
\usage{
plotPromoterDMRs(x, tx, CpGislands, outFile, zoom = 3000,
  proteinOnly = FALSE, CpGi = TRUE, mc.cores = 1, hub, session = NULL)
}
\arguments{
  \item{x}{\code{GRanges} of regions to screenshot}

  \item{tx}{\code{GRanges} of transcript expression}

  \item{CpGislands}{\code{GRanges} of CpG island positions}

  \item{outFile}{Filename of the output pdf}

  \item{zoom}{The distance to pad either side of the
  regions of interest by}

  \item{proteinOnly}{Whether to restrict closest TSS
  plotting to protein coding transcripts only, or to
  include non-coding transcripts}

  \item{CpGi}{Whether to include the closest CpG island in
  the screenshot}

  \item{mc.cores}{No of threads to use to take screenshots
  - too many will break UCSC/this function}

  \item{hub}{URL of UCSC track hub to load}

  \item{session}{URL of the UCSC session to load
  (optional)}
}
\description{
Takes UCSC screenshots which contain a region of interest,
its closest TSS and (optionally) its closest CpG island
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

