% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyalpha.R
\name{prettyalpha}
\alias{prettyalpha}
\title{Pretty Printing of psych::alpha() Results}
\usage{
prettyalpha(alphaobj, ...)
}
\arguments{
\item{alphaobj}{Summary object from executing psych::alpha().}

\item{...}{Additional arguments passed onto \code{kable()}.}
}
\value{

}
\description{
Results from \code{psych::alpha()} get messed up when \code{alpha()} is called within a for loop within an RMarkdown document with the chunk option \code{results = "asis"}.
This function is an alternative to \code{alpha()} that can print out a pretty output within such a loop.
}
\examples{
To be added...
}
