% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettydescribeBy.R
\name{prettydescribeBy}
\alias{prettydescribeBy}
\title{Pretty formatting for psych::describeBy()}
\usage{
prettydescribeBy(desc_out, y)
}
\arguments{
\item{desc_out}{Output from describeBy()}

\item{y}{Variable of interest}
}
\value{
Dataframe with outputs from psych::describeBy()
}
\description{
Format the output of describeBy() as columns for the statistics of different groups
}
\examples{
mydata <- data.frame(x = c(rep("Control", 50), rep("Treatment", 50)), y = c(1:50,101:150))
psych::describeBy(mydata, group = "x") \%>\% prettydescribeBy(y = "y")
}
