% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettylavaan.R
\name{prettylavaan}
\alias{prettylavaan}
\title{Pretty Printing of Lavaan Results}
\usage{
prettylavaan(fitobj, output_format = "asis")
}
\arguments{
\item{fitobj}{A lavaan fit object.}

\item{output_format}{Output format, currently only "asis" is implemented.}
}
\value{

}
\description{
Results from \code{lavaan} get messed up when \code{summary()} is called within a for loop within an RMarkdown document with the chunk option \code{results = "asis"}.
This function is an alternative to \code{summary()} that can print out a pretty output within such a loop.
}
\examples{
TODO
}
