% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettytable.R
\name{prettytable}
\alias{prettytable}
\title{Printing 1 & 2 Way Frequency Tables With Relative Proportions}
\usage{
prettytable(mytable, margin = 1)
}
\arguments{
\item{margin}{Defaults to 1, calculate row-wise proportions. Change to 2 to calculate column-wise proportions.}

\item{table}{A table object from \code{table()}.}
}
\value{
Returns a table with relative proportions/percentages appended to raw counts.
}
\description{
Automatically append relative proportions/percentages to raw counts calculated from \code{table()}. Currently only works for 1 and 2 way frequency tables.
}
\examples{
mydata <- data.frame(x1 = rep(c("Male","Female","Male"), 100), x2 = rep(c("High","Medium","Low"),100))
# two way frequency table
a <- table(mydata$x1,mydata$x2)
prettytable(a)
# one way frequency table
a <- table(mydata$x1)
prettytable(a)
}
