% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob2odd.R
\name{prob2odd}
\alias{prob2odd}
\title{Convert Probabilities To Odds}
\usage{
prob2odd(p)
}
\arguments{
\item{p}{Probability value to be converted}
}
\value{
Return the odds
}
\description{
Convert probabilities to odds using the formula of odds = p/(1-p)
}
\examples{
p <- c(0.3,0.4,0.5)
prob2odd(p)
}
