% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regionalAmpliconAnalysis}
\alias{regionalAmpliconAnalysis}
\title{regionalAmpliconAnalysis}
\usage{
regionalAmpliconAnalysis(x, grouping, minCov = 50)
}
\arguments{
  \item{x}{Result returned by \link{ampliconAnalysis}}

  \item{grouping}{\code{character} vector of region names
  by which amplicons are grouped together}

  \item{minCov}{Minumum sequencing coverage required to
  report methylation/conversion estimates}
}
\value{
List
}
\description{
Condenses the analysis of individual amplicons to those
grouped in regions
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

