% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screenshotUCSC}
\alias{screenshotUCSC}
\title{screenshotUCSC}
\usage{
screenshotUCSC(url, hubfile, chr, start, end, filename, session = NULL)
}
\arguments{
  \item{url}{UCSC URL to use, sets genome and screenshot
  resolution}

  \item{hubfile}{URL of UCSC track hub to load}

  \item{chr}{Chromosome to screenshot}

  \item{start}{Start of the region to screenshot}

  \item{end}{End of the region to screenshot}

  \item{filename}{Filename of the output pdf}

  \item{session}{URL of the UCSC session to load
  (optional)}
}
\value{
Called for the side effect of downloading a pdf screenshot,
returns invisible debugging output
}
\description{
Load a UCSC track hub and save a screenshot of a given
region as a pdf
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

