% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seqHeatmaps}
\alias{seqHeatmaps}
\title{seqHeatmaps}
\usage{
seqHeatmaps(bamfiles, anno, up = 2000, down = 2000, freq = 20,
  s.width = 200, nc = length(anno), nr = length(bamfiles),
  cols = colorRampPalette(c("white", "red"))(100), zlim = c(0, 1),
  n.sample = 2000)
}
\arguments{
  \item{bamfiles}{Paths to BAM files with names}

  \item{anno}{\code{GRangesList} or \code{GRanges} of
  regions to sample around}

  \item{up}{How far upstream to sample}

  \item{down}{How far downstream to sample}

  \item{freq}{How many basepairs apart to sample}

  \item{s.width}{How many basepairs to smooth the data by}

  \item{nc}{Number of columns to plot per page}

  \item{nr}{Number of rows to plot per page}

  \item{cols}{colors to use in the heatmap}

  \item{zlim}{limit to plot in the heatmap, defaults to 0-1
  read per million}

  \item{n.sample}{Number of regions from each anno to
  randomly sample for plotting, NULL disables the sampling}
}
\value{
Plots a set of heatmaps and invisibly returns the plotted
data
}
\description{
Plot a grid of heatmaps of ChIP-seq signals around regions
of interest
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

