% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffleBed}
\alias{shuffleBed}
\title{shuffleBed}
\usage{
shuffleBed(x)
}
\arguments{
  \item{x}{Set of ranges to be shuffled}
}
\value{
A shuffled \code{GRanges} of the same length as \code{x}
}
\description{
A hack to use bedtools shuffleBed to shuffle ranges across
a genome
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

