% Generated by roxygen2 (4.0.1): do not edit by hand
\name{slidingMethRatios}
\alias{slidingMethRatios}
\title{slidingMethRatios}
\usage{
slidingMethRatios(x, samples, n = 5, minCov = 5)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{samples}{\code{data.frame} describing the samples
  to calculate ratios for}

  \item{n}{Number of adjacent CpG sites to average}

  \item{minCov}{minimum amount of coverage to calculate a
  ratio}
}
\value{
\code{GRanges} of methylation ratio means across \code{n}
CpG sites
}
\description{
Calculates methylation ratios means across a sliding window
of a number of CpG sites
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

