% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coverageBarplot.R
\name{tabulateCoverage}
\alias{tabulateCoverage}
\title{tabulateCoverage}
\usage{
tabulateCoverage(regions, background, annotations)
}
\arguments{
  \item{regions}{\code{GRangesList} of query regions, for
  example DMRs}

  \item{background}{\code{GRanges} of interrogated genomic
  background}

  \item{annotations}{\code{GRangesList} of subject regions,
  for example ChromHMM annotations}
}
\value{
A \code{list} containing the elements:
\item{coverage}{Bases intersected between ranges of
\code{regions} and \code{annotations}}
\item{ratios}{Percentage of genome covered by each
intersection} \item{observed/expected}{Enrichment ratio of
\code{regions} regions taking into account the
\code{background}}
}
\description{
Tabulate how much of the genome is each annotation
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

