% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tissueSpecificDMRs.R
\name{tissueSpecificDMRs}
\alias{tissueSpecificDMRs}
\title{tissueSpecificDMRs}
\usage{
tissueSpecificDMRs(x, samples, k = 3, maxCov = 200, p.cutoff = 0.01,
  minCpGs = 5, minWidth = 100, minCpGdensity = 5, minSd = 0.1,
  mc.cores = 1)
}
\arguments{
  \item{x}{CpG-site level methylation data}

  \item{samples}{\code{data.frame} describing the samples
  to process}

  \item{k}{Number of adjacent CpG sites to calculate chi-sq
  statistics upon}

  \item{maxCov}{Maximum sequencing depth allowable at a
  single CpG site; essential for removing repeat regions}

  \item{p.cutoff}{Chi-sq BY adjusted p-value for a CpG site
  to be considered significant}

  \item{minCpGs}{Minimum number of adjacent CpG sites below
  \code{p.cutoff} for a region to be considered
  significant}

  \item{minWidth}{Minimum size in basepairs for a tsDMR}

  \item{minCpGdensity}{Minimum number of CpG sites per
  kilobase in a tsDMR}

  \item{minSd}{Minimum standard deviation of methylaion
  ratios of a tsDMR across the tissues}

  \item{mc.cores}{Number of cores to use}
}
\value{
\code{GRanges} of significant regions, with
mean-methyaltion per sample and chisq value in the values()
}
\description{
Find regions of tissue specific methylation differences
using a sliding chi-squared test. Method adapted from
"Adult tissue methylomes harbor epigenetic memory at
embryonic enhancers", Hon et al, Nature Genetics 2013
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

