% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{unstrand}
\alias{unstrand}
\title{unstrand}
\usage{
unstrand(x)
}
\arguments{
  \item{x}{A \code{GRanges} objext}
}
\value{
\code{x} with strand set to '*'
}
\description{
Remove strand data from a GRanges object
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

