% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{unvalue}
\alias{unvalue}
\title{unvalue}
\usage{
unvalue(x)
}
\arguments{
  \item{x}{A \code{GRanges} objext}
}
\value{
\code{x} with no element metadata
}
\description{
Remove element metadata from a GRanges object
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

