% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.style.R
\name{use.style}
\alias{use.style}
\title{Applying \code{aaRon.css} to Rmarkdown Documents}
\usage{
use.style(mode = "work")
}
\arguments{
\item{mode}{Defaults to "work", can take the option of "casual". Switches between CSS themes for work and casual use. The casual theme is slightly more flashy.}
}
\value{

}
\description{
Function to apply cascading stylesheets from \code{aaRon} to any Rmarkdown document. Also accessible through the templates provided in the package.
}
\examples{
aaRon::use.style() # include into any R code chunk
aaRon::use.style(mode = "casual")
}
