% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{}
\usage{
vlookup(data, lookup_table, data_col, lookup_col, lookup_value)
}
\arguments{
\item{data}{WIP}

\item{lookup_table}{WIP}

\item{data_col}{WIP}

\item{lookup_col}{WIP}

\item{lookup_value}{WIP}
}
\value{

}
\description{

}
\examples{
mydata <- data.frame(Type = rep(1:3,5), Rand = 4:6)
mytable <- data.frame(Type = 1:3, Magnitude = c("Low","Medium","High"))
vlookup(data = mydata,
        lookup_table = mytable,
        data_col = "Type",
        lookup_col = "Type",
        lookup_value = "Magnitude")
}
