% Generated by roxygen2 (4.0.2): do not edit by hand
\name{weightedMethylation}
\alias{weightedMethylation}
\title{weightedMethylation}
\usage{
weightedMethylation(x, regions, w)
}
\arguments{
  \item{x}{\code{GRanges} of CpG-site level methylation
  ratios}

  \item{regions}{\code{GRanges} of genomic regions to be
  plotted}

  \item{w}{\code{vector} of CpG-site level weights
  (optional)}
}
\value{
\code{matrix} of weighted methylation values
}
\description{
Calculate a per sample weighted mean of CpG-site level
methylation data
}

