% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ampliconAnalysis}
\alias{ampliconAnalysis}
\title{ampliconAnalysis}
\usage{
ampliconAnalysis(amplicon_file, bams, genome, minCov = 50, min.map.q = 40,
  mc.cores = 1)
}
\arguments{
  \item{amplicon_file}{Filename of the \code{csv} to read
  in amplicon data from}

  \item{genome}{\code{BSgenome} to map the amplicons to}

  \item{bams}{A named \code{character} vector of bams to be
  analysed for the supplied amplicons}

  \item{minCov}{Minumum sequencing coverage required to
  report methylation/conversion estimates}

  \item{min.map.q}{Minumum mapping quality for reads to be
  included in the analysis, set to 0 for no filtering}

  \item{mc.cores}{Number of cores to use during processing}
}
\value{
List
}
\description{
Start to finish analysis of MiSeq bisulfite amplicon
sequencing data
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

