% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binarizeFromPeaks.R
\name{binarizeFromPeaks}
\alias{binarizeFromPeaks}
\title{binarizeFromPeaks}
\usage{
binarizeFromPeaks(peaks, outdir, resolution = 200, seqlen)
}
\arguments{
  \item{peaks}{\code{data.frame} with column names
  \code{sample}, \code{mark}, \code{bed}}

  \item{outdir}{Path to directory to output binary files
  to}

  \item{resolution}{Resolution in basepairs to bin the
  genome into}

  \item{seqlen}{Named \code{character} vector of chromosome
  lengths}
}
\value{
Called for side effect of creating binary files
}
\description{
Converts bed files of peaks (eg from MACS or PeakRanger) to
binary files for ChromHMM
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

