% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coverageRatio}
\alias{coverageRatio}
\title{coverageRatio}
\usage{
coverageRatio(query, subject, ratio = TRUE)
}
\arguments{
  \item{query}{query ranges}

  \item{subject}{subject ranges}

  \item{ratio}{If \code{TRUE} return the ratio, if
  \code{FALSE} return the number of basepairs overlap}
}
\value{
A \code{vector} of the same length as \code{subject}
containing the calculated overlap
}
\description{
Calculate the amount of each \code{subject} ranges overlap
with \code{query} ranges
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

