% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct2num.R
\name{fct2num}
\alias{fct2num}
\title{fct2num()}
\usage{
fct2num(factor_vector, start = 1, print = "message")
}
\arguments{
\item{factor_vector}{A vector that is a factor.}

\item{start}{Defaults to 1. The starting value of the resultant numeric vector. The value inputted for \code{start} will be the lowest numeric value for the lowest factor level.}

\item{print}{Defaults to "message". Determines if the text messages of this function uses \code{cat()} or \code{message()}. This affects the behaviour in specific Rmarkdown chunks.}
}
\value{

}
\description{
Converts a factor to a numeric vector while adjusting for the starting value of the numeric vector.
Useful for converting from categorical responses (e.g. "Agree","Disagree") to numeric (1,2).
}
\examples{
a <- factor(c("Agree","Disagree","Agree","Agree","Disagree"))
fct2num(a)
fct2num(a, start = 0) # if likert scales begin at 0
}
