% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findDMVs}
\alias{findDMVs}
\title{findDMVs}
\usage{
findDMVs(x, samples, minSize = 5000, wsize = 1000, step = 100,
  cutoff = 0.15, minCov = 5)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{samples}{\code{data.frame} describing the samples
  to discover DMVs for}

  \item{minSize}{Minimum number of basepairs with
  methylation below \code{cutoff} for a region to be called
  as a DMV}

  \item{wsize}{Window size to take weighted means within
  across the genome}

  \item{step}{Distance to step windows across the genome
  by}

  \item{cutoff}{Maximum weighted mean of a window to
  contribute to a DMV}

  \item{minCov}{Minumum sequencing coverage for a CpG site
  to contribute to the weighted mean calculation}
}
\value{
\code{GRangesList} of DMVs found in each sample

importFrom GenomicRanges values seqlengths seqnames
GRangesList reduce width split importFrom Repitools
genomeBlocks importFrom parallel multicore
}
\description{
Finds DNA methylation valleys using the procedure described
in Hovestaft 2014
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

