% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findNDRs.R
\name{findNDRs}
\alias{findNDRs}
\title{findNDRs}
\usage{
findNDRs(x, samples, p.cutoff = 15, windowWidth = 100, windowBy = 20,
  minSize = 140)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data at GCH
  sites}

  \item{samples}{\code{data.frame} describing the samples
  to discover NDRs for}

  \item{p.cutoff}{Minimum -log10 p.value for a window to be
  considered significant}

  \item{windowWidth}{Size of each windoe}

  \item{windowBy}{Distance between the starts of
  consecutive windows}

  \item{minSize}{Minimum size of an NDR when overlapping
  significant windows are pooled}
}
\value{
\code{GRangesList} of NDRs found in each sample with the
mean pvalue for each region
}
\description{
Find NDRs in NOMe-seq data using a sliding window
chi-squared test versus whole genome background
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

