% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grangesVenn}
\alias{grangesVenn}
\title{grangesVenn}
\usage{
grangesVenn(x, plot = TRUE, ...)
}
\arguments{
  \item{x}{\code{GRangesList} of interest}

  \item{plot}{Whether to plot after calculating the}

  \item{...}{Additional parameters passed on to
  vennDiagram}
}
\value{
A list containing the pooled and reduced ranges, a boolean
matrix of whether each sample in \code{x} overlaps each
\code{Ranges} and the vennCounts
}
\description{
Creates a table of overlapping \code{GRanges} from a
\code{GRangesList}, vennCounts and optionally plots the
first (up to) five
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

