% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overlapBoundaries}
\alias{overlapBoundaries}
\title{overlapBoundaries}
\usage{
overlapBoundaries(query, subject, distance = 10000, boundary = c("both",
  "start", "end"))
}
\arguments{
  \item{query}{\code{GRanges} of query regions}

  \item{subject}{\code{GRanges} of subject regions}

  \item{boundary}{Whether to interrogate the start, end or
  both boundaries of the \code{query} range}

  \item{distance}{Number of basepairs up and downstream of
  \code{subject} regions to interrogate}
}
\value{
Positions where \code{subject} ranges overlap the
boundaries of \code{query} ranges
}
\description{
Returns
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

