% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overlapRegions}
\alias{overlapRegions}
\title{overlapRegions}
\usage{
overlapRegions(query, subject)
}
\arguments{
  \item{query}{\code{GRanges} of query regions}

  \item{subject}{\code{GRanges} of subject regions (must be
  of width 1 and unstranded)}
}
\value{
\code{data.frame} of index and position
}
\description{
Returns the relative position of \code{subject} sites
within \code{query} regions
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

