% Generated by roxygen2 (4.0.1): do not edit by hand
\name{poolGroups}
\alias{poolGroups}
\title{poolGroups}
\usage{
poolGroups(x, samples)
}
\arguments{
  \item{x}{\code{GRanges} of methylation count data}

  \item{samples}{\code{data.frame} describing the samples
  and their grouping}
}
\value{
A \code{list} containing the elements \code{CpGs} and
\code{samples} with the groups pooled
}
\description{
Pools methylation count data of samples from the same group
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

