% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyalpha.R
\name{prettyalpha}
\alias{prettyalpha}
\title{Pretty Printing of psych::alpha() Results}
\usage{
prettyalpha(alphaobj, dp = 2, align = "c", ...)
}
\arguments{
\item{alphaobj}{Summary object from executing psych::alpha().}

\item{dp}{Defaults to 3. Number of decimal points for all numeric values.}

\item{align}{Defaults to "c", which stands for centered. Adjusts the alignment of text within table cells. Works the same way as \code{kable()}.}

\item{...}{Additional arguments passed onto \code{kable()}.}
}
\value{

}
\description{
Results from \code{psych::alpha()} get messed up when \code{alpha()} is called within a \code{for} loop within an R Markdown document with the chunk option \code{results = "asis"}.
This function is an alternative to \code{alpha()} that can print out a pretty output in such situations.
}
\examples{
myalpha <- alpha(HolzingerSwineford1939[,7:15])
prettyalpha(myalpha)
}
