% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettylavaan.R
\name{prettylavaan}
\alias{prettylavaan}
\title{Pretty Printing of Lavaan Results}
\usage{
prettylavaan(
  fitobj,
  output_format = "asis",
  robust = FALSE,
  modindice.nrow = 10,
  param.type = c("=~", "~~", "~"),
  dp = 3,
  align = "c",
  ...
)
}
\arguments{
\item{fitobj}{A lavaan fit object.}

\item{output_format}{Output format, default is "asis" for use with \code{results = "asis"} chunks. Use "datatable" for an interactive parameter table. Or "kableExtra" to generate a list of kable tables that can be independently piped to kableExtra functions.}

\item{robust}{Defaults to \code{FALSE}, set to \code{TRUE} to print out scaled and robust fit indicators.}

\item{modindice.nrow}{Defaults to 10. Number of rows to display for modification indices.}

\item{param.type}{Defaults to only show path coefficients (~), factor loadings (=~) and covariances (~~). Use (|) for thresholds if WLS estimators were used.}

\item{dp}{Defaults to 3. Number of decimal points for all numeric values.}

\item{align}{Defaults to "c", which stands for centered. Adjusts the alignment of text within table cells. Works the same way as \code{kable()}.}

\item{...}{Additional arguments passed onto \code{kable()} or \code{datatable} depending on \code{output_format}.}
}
\value{

}
\description{
Results from \code{lavaan} get messed up when \code{summary()} is called within a for loop within an RMarkdown document with the chunk option \code{results = "asis"}.
This function is an alternative to \code{summary()} that can print out a pretty output within such a loop.
}
\examples{
library(lavaan)
# the famous Holzinger and Swineford (1939) example
HS.model <-  "visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9"
fit <- cfa(HS.model, data = HolzingerSwineford1939)
prettylavaan(fit, output_format = "datatable")

# using robust estimators
robustfit <- cfa(HS.model, data = HolzingerSwineford1939, estimator = "MLM")
prettylavaan(robustfit, output_format = "datatable")

# request for robust fit indices
prettylavaan(robustfit, output_format = "datatable", robust = TRUE)

# for piping to kableExtra for further editing
# note: library(kableExtra) may mess up formatting of normal kable tables in the same Rmd document
# refer to https://github.com/haozhu233/kableExtra/issues/265
# format = "html" is required to work with kableExtra
library(kableExtra)
mylist <- prettylavaan(robustfit, output_format = "kableExtra", format = "html", robust = TRUE)
mylist$Param \%>\% kable_styling(font_size = 9)
}
