% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettylavaan.R
\name{prettylavaan}
\alias{prettylavaan}
\title{Pretty Printing of Lavaan Results}
\usage{
prettylavaan(fitobj, output_format = "asis")
}
\arguments{
\item{fitobj}{A lavaan fit object.}

\item{output_format}{Output format, either "asis" for use with \code{results = "asis"} chunks. Or "datatable" for general html output.}
}
\value{

}
\description{
Results from \code{lavaan} get messed up when \code{summary()} is called within a for loop within an RMarkdown document with the chunk option \code{results = "asis"}.
This function is an alternative to \code{summary()} that can print out a pretty output within such a loop.
}
\examples{
## The famous Holzinger and Swineford (1939) example
HS.model <-  "visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9"
fit <- cfa(HS.model, data = HolzingerSwineford1939)
prettylavaan(fit, output_format = "datatable")
}
