% Generated by roxygen2 (4.0.0): do not edit by hand
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(x)
}
\arguments{
  \item{x}{Set of ranges to be shuffled}
}
\value{
A shuffled \code{GRanges} of the same length as \code{x}
}
\description{
Shuffle a set of \code{GRanges} across an available genome
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

