% Generated by roxygen2 (4.0.0): do not edit by hand
\name{strip}
\alias{strip}
\title{strip}
\usage{
strip(x, genome, chrs = 1:24)
}
\arguments{
  \item{x}{A \code{GRanges} objext}

  \item{genome}{\code{BSgenome} object}

  \item{chrs}{Chromosomes to limit \code{x} to}
}
\value{
Reduced \code{GRanges} of restricted chromosomes, with no
strand or element metadata
}
\description{
Cleans up a \code{GRanges} object
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

