% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{strip}
\alias{strip}
\title{strip}
\usage{
strip(x)
}
\arguments{
  \item{x}{A \code{GRanges} objext}
}
\value{
Reduced \code{GRanges} with no strand or element metadata
}
\description{
Cleans up a \code{GRanges} object
}
\author{
Aaron Statham <a.statham@garvan.org.au>
}

