% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.style.R
\name{use.style}
\alias{use.style}
\title{Applying \code{aaRon.css} to Rmarkdown Documents}
\usage{
use.style()
}
\value{

}
\description{
Function to apply cascading stylesheets from \code{aaRon} to any Rmarkdown document.
}
\examples{
aaRon::use.style() # include into any R code chunk
}
